[CLEANUP] Applied pull request changes (https://github.com/irvinlim/acf-recaptcha/pull/14)
