+++ lib/ReCaptcha/ReCaptcha.php
@@ -70,7 +70,7 @@ class ReCaptcha
         if (!is_null($requestMethod)) {
             $this->requestMethod = $requestMethod;
         } else {
-            $this->requestMethod = new RequestMethod\WP();
+            $this->requestMethod = new RequestMethod\Post();
         }
     }
