--- lib/ReCaptcha/RequestMethod/WP.php
@@ -1,63 +1,0 @@
-<?php
-/**
- * This is a PHP library that handles calling reCAPTCHA.
- *
- * @copyright Copyright (c) 2015, Google Inc.
- * @link      http://www.google.com/recaptcha
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-namespace ReCaptcha\RequestMethod;
-
-use ReCaptcha\RequestMethod;
-use ReCaptcha\RequestParameters;
-
-/**
- * Sends POST requests to the reCAPTCHA service using the WP remote POST method
- */
-class WP implements RequestMethod
-{
-    /**
-     * URL to which requests are POSTed.
-     * @const string
-     */
-    const SITE_VERIFY_URL = 'https://www.google.com/recaptcha/api/siteverify';
-
-    /**
-     * Submit the POST request with the specified parameters.
-     *
-     * @param RequestParameters $params Request parameters
-     * @return string Body of the reCAPTCHA response
-     */
-    public function submit(RequestParameters $params)
-    {
-    	$response = wp_remote_post( self::SITE_VERIFY_URL, array(
-    			'method' => 'POST',
-    			'body'   => $params->toQueryString(),
-    	) );
-    	
-    	if ( ! is_wp_error( $response ) ) {
-    		return wp_remote_retrieve_body( $response );
-    	}
-    	
-    	// Alltough we try to decode json we will throw the error
-    	throw $response->get_error_message();
-    }
-}
